Working delay in sim
